Revert "Merge pull request #351 from mtane0412/refactor/create"
This reverts commit a5d738a39c33d3c5f0a72d6fad216d8ab08b9a23, reversing
changes made to 66ba9759325c169c2915998d3665d89b75850fb4.
