+++ client-admin/app/create/page.tsx
@@ -1,84 +1,269 @@
 "use client";
 
+import { initialLabellingPrompt } from "@/app/create/initialLabellingPrompt";
+import { mergeLabellingPrompt } from "@/app/create/mergeLabellingPrompt";
+import { overviewPrompt } from "@/app/create/overviewPrompt";
+import { type CsvData, parseCsv } from "@/app/create/parseCsv";
 import { Header } from "@/components/Header";
+import { Checkbox } from "@/components/ui/checkbox";
+import {
+  FileUploadDropzone,
+  FileUploadList,
+  FileUploadRoot,
+} from "@/components/ui/file-upload";
 import { toaster } from "@/components/ui/toaster";
 import {
+  Alert,
   Box,
   Button,
   Field,
   HStack,
   Heading,
+  Input,
+  NativeSelect,
   Presence,
+  Stack,
   Tabs,
+  Text,
+  Textarea,
   VStack,
   useDisclosure,
 } from "@chakra-ui/react";
+import { ChevronRightIcon, DownloadIcon } from "lucide-react";
+import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
-import { createReport } from "./api/report";
-import { AISettingsSection } from "./components/AISettingsSection";
-import { BasicInfoSection } from "./components/BasicInfoSection";
-import { CsvFileTab } from "./components/CsvFileTab";
-import { SpreadsheetTab } from "./components/SpreadsheetTab";
-import { WarningSection } from "./components/WarningSection";
-import { useAISettings } from "./hooks/useAISettings";
-import { useBasicInfo } from "./hooks/useBasicInfo";
-import { useClusterSettings } from "./hooks/useClusterSettings";
-import { useInputData } from "./hooks/useInputData";
-import { usePromptSettings } from "./hooks/usePromptSettings";
-import { CsvData, parseCsv } from "./parseCsv";
-import { showErrorToast } from "./utils/error-handler";
-import { validateFormValues } from "./utils/validation";
+import { v4 } from "uuid";
+import { extractionPrompt } from "./extractionPrompt";
 
-/**
- * レポート作成ページ
- */
+interface SpreadsheetComment {
+  id?: string;
+  comment: string;
+  source?: string | null;
+  url?: string | null;
+}
+
 export default function Page() {
   const router = useRouter();
   const { open, onToggle } = useDisclosure();
   const [loading, setLoading] = useState<boolean>(false);
+  const [input, setInput] = useState<string>(v4());
+  const [importedId, setImportedId] = useState<string>("");
+  const [question, setQuestion] = useState<string>("");
+  const [intro, setIntro] = useState<string>("");
+  const [csv, setCsv] = useState<File | null>(null);
+  const [inputType, setInputType] = useState<"file" | "spreadsheet">("file");
+  const [spreadsheetUrl, setSpreadsheetUrl] = useState<string>("");
+  const [spreadsheetImported, setSpreadsheetImported] =
+    useState<boolean>(false);
+  const [spreadsheetLoading, setSpreadsheetLoading] = useState<boolean>(false);
+  const [spreadsheetData, setSpreadsheetData] = useState<SpreadsheetComment[]>(
+    [],
+  );
+  const [model, setModel] = useState<string>("gpt-4o-mini");
+  const [clusterLv1, setClusterLv1] = useState<number>(5);
+  const [clusterLv2, setClusterLv2] = useState<number>(50);
+  const [workers, setWorkers] = useState<number>(30);
+  const [extraction, setExtraction] = useState<string>(extractionPrompt);
+  const [initialLabelling, setInitialLabelling] = useState<string>(
+    initialLabellingPrompt,
+  );
+  const [mergeLabelling, setMergeLabelling] =
+    useState<string>(mergeLabellingPrompt);
+  const [overview, setOverview] = useState<string>(overviewPrompt);
+  const [isPubcomMode, setIsPubcomMode] = useState<boolean>(true);
+  const [csvColumns, setCsvColumns] = useState<string[]>([]);
+  const [selectedCommentColumn, setSelectedCommentColumn] =
+    useState<string>("");
+  const [recommendedClusters, setRecommendedClusters] = useState<{
+    lv1: number;
+    lv2: number;
+  } | null>(null);
+  const [autoAdjusted, setAutoAdjusted] = useState<boolean>(false);
 
-  // カスタムフックの使用
-  const basicInfo = useBasicInfo();
-  const clusterSettings = useClusterSettings();
-  const promptSettings = usePromptSettings();
-  const aiSettings = useAISettings();
-  const inputData = useInputData(clusterSettings.setRecommended);
-
-  /**
-   * タブ切り替え時の処理
-   */
-  const handleTabValueChange = (details: { value: string }) => {
-    inputData.setInputType(details.value as any);
+  // IDのバリデーション関数
+  const isValidId = (id: string): boolean => {
+    return /^[a-z0-9](?:[a-z0-9-]*[a-z0-9])?$/.test(id) && id.length <= 255;
   };
 
-  /**
-   * レポート作成の送信
-   */
-  const onSubmit = async () => {
-    setLoading(true);
+  // IDのバリデーション結果を状態として持つ
+  const [isIdValid, setIsIdValid] = useState<boolean>(true);
 
-    // フォーム入力値のバリデーション
-    const validation = validateFormValues({
-      input: basicInfo.input,
-      question: basicInfo.question,
-      intro: basicInfo.intro,
-      clusterLv1: clusterSettings.clusterLv1,
-      clusterLv2: clusterSettings.clusterLv2,
-      model: aiSettings.model,
-      extractionPrompt: promptSettings.extraction,
-      inputType: inputData.inputType,
-      csv: inputData.csv,
-      spreadsheetImported: inputData.spreadsheetImported,
-      selectedCommentColumn: inputData.selectedCommentColumn,
-      csvColumns: inputData.csvColumns,
-    });
+  // 入力変更時にバリデーションを実行
+  const handleIdChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const newId = e.target.value;
+    setInput(newId);
+    setIsIdValid(isValidId(newId));
+  };
 
-    if (!validation.isValid) {
+  const calculateRecommendedClusters = (commentCount: number) => {
+    const lv1 = Math.max(2, Math.min(20, Math.round(Math.cbrt(commentCount))));
+    const lv2 = Math.max(2, Math.min(1000, Math.round(lv1 * lv1)));
+
+    return { lv1, lv2 };
+  };
+
+  const canImport = spreadsheetUrl.trim() && isIdValid && !spreadsheetImported;
+
+  async function importSpreadsheet() {
+    if (!canImport) {
       toaster.create({
         type: "error",
         title: "入力エラー",
-        description: validation.errorMessage,
+        description: !isIdValid
+          ? "IDは英小文字、数字、ハイフンのみ使用できます"
+          : "スプレッドシートのURLを入力してください",
+      });
+      return;
+    }
+
+    setSpreadsheetLoading(true);
+    try {
+      const response = await fetch(
+        `${process.env.NEXT_PUBLIC_API_BASEPATH}/admin/spreadsheet/import`,
+        {
+          method: "POST",
+          headers: {
+            "x-api-key": process.env.NEXT_PUBLIC_ADMIN_API_KEY || "",
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify({
+            url: spreadsheetUrl,
+            file_name: input,
+          }),
+        },
+      );
+
+      if (!response.ok) {
+        const errorData = await response.json();
+        throw new Error(errorData.detail || "不明なエラーが発生しました");
+      }
+
+      await response.json();
+
+      setImportedId(input);
+
+      // スプレッドシートのデータを取得
+      const commentResponse = await fetch(
+        `${process.env.NEXT_PUBLIC_API_BASEPATH}/admin/spreadsheet/data/${input}`,
+        {
+          headers: {
+            "x-api-key": process.env.NEXT_PUBLIC_ADMIN_API_KEY || "",
+          },
+        },
+      );
+
+      if (!commentResponse.ok) {
+        throw new Error("スプレッドシートデータの取得に失敗しました");
+      }
+
+      const commentData = await commentResponse.json();
+      setSpreadsheetData(commentData.comments);
+
+      if (commentData.comments.length > 0) {
+        const columns = Object.keys(commentData.comments[0]);
+        setCsvColumns(columns);
+        if (columns.includes("comment")) {
+          setSelectedCommentColumn("comment");
+        }
+      }
+
+      if (commentData.comments.length > 0) {
+        setCsvColumns(Object.keys(commentData.comments[0]));
+      }
+
+      const recommended = calculateRecommendedClusters(commentData.comments.length);
+      setRecommendedClusters(recommended);
+      // 推奨クラスタ数を自動的に適用
+      setClusterLv1(recommended.lv1);
+      setClusterLv2(recommended.lv2);
+
+      toaster.create({
+        type: "success",
+        title: "成功",
+        description: `スプレッドシートのデータ ${commentData.comments.length} 件を取得しました`,
+      });
+      setSpreadsheetImported(true);
+    } catch (e) {
+      console.error(e);
+
+      // エラーメッセージを解析して、より適切な短いメッセージに変換
+      const errorMessage =
+        e instanceof Error ? e.message : "不明なエラーが発生しました";
+      let displayErrorMessage = "";
+
+      // URLやアクセス権限のエラー
+      if (
+        errorMessage.includes("Unauthorized") ||
+        errorMessage.includes("401")
+      ) {
+        displayErrorMessage =
+          "スプレッドシートへのアクセス権限がありません。公開設定を確認してください。";
+      }
+      // 存在しないシートなど
+      else if (
+        errorMessage.includes("404") ||
+        errorMessage.includes("Not Found")
+      ) {
+        displayErrorMessage =
+          "スプレッドシートが見つかりません。URLを確認してください。";
+      }
+      // スプレッドシート形式の問題
+      else if (
+        errorMessage.includes("comment") ||
+        errorMessage.includes("カラム")
+      ) {
+        displayErrorMessage =
+          "スプレッドシートの形式が正しくありません。commentカラムが必要です。";
+      }
+      // その他のエラー
+      else {
+        displayErrorMessage =
+          "スプレッドシートの取得に失敗しました。URLと公開設定を確認してください。";
+      }
+
+      toaster.create({
+        type: "error",
+        title: "データ取得エラー",
+        description: displayErrorMessage,
+      });
+      setSpreadsheetImported(false);
+    } finally {
+      setSpreadsheetLoading(false);
+    }
+  }
+
+  async function onSubmit() {
+    setLoading(true);
+
+    const commonCheck = [
+      /^[a-z0-9](?:[a-z0-9-]*[a-z0-9])?$/.test(input),
+      question.length > 0,
+      intro.length > 0,
+      clusterLv1 > 0,
+      clusterLv2 > 0,
+      model.length > 0,
+      extraction.length > 0,
+    ].every(Boolean);
+
+    const sourceCheck =
+      (inputType === "file" && !!csv) ||
+      (inputType === "spreadsheet" && spreadsheetImported);
+
+    if (!commonCheck || !sourceCheck) {
+      toaster.create({
+        type: "error",
+        title: "入力エラー",
+        description: "全ての項目が入力されているか確認してください",
+      });
+      setLoading(false);
+      return;
+    }
+    if (csvColumns.length > 0 && !selectedCommentColumn) {
+      toaster.create({
+        type: "error",
+        title: "カラム未選択",
+        description: "コメントカラムを選択してください",
       });
       setLoading(false);
       return;
@@ -86,20 +271,19 @@ export default function Page() {
 
     let comments: CsvData[] = [];
     try {
-      if (inputData.inputType === "file" && inputData.csv) {
-        const parsed = await parseCsv(inputData.csv);
+      if (inputType === "file" && csv) {
+        const parsed = await parseCsv(csv);
         comments = parsed.map((row, index) => ({
           id: `csv-${index + 1}`,
           comment: (row as unknown as Record<string, unknown>)[
-            inputData.selectedCommentColumn
+            selectedCommentColumn
           ] as string,
           source: null,
           url: null,
         }));
-        
-        if (comments.length < clusterSettings.clusterLv2) {
+        if (comments.length < clusterLv2) {
           const confirmProceed = window.confirm(
-            `csvファイルの行数 (${comments.length}) が設定された意見グループ数 (${clusterSettings.clusterLv2}) を下回っています。このまま続けますか？
+            `csvファイルの行数 (${comments.length}) が設定された意見グループ数 (${clusterLv2}) を下回っています。このまま続けますか？
     \n※コメントから抽出される意見が設定された意見グループ数に満たない場合、処理中にエラーになる可能性があります（一つのコメントから複数の意見が抽出されることもあるため、問題ない場合もあります）。
     \n意見グループ数を変更する場合は、「AI詳細設定」を開いてください。`,
           );
@@ -108,11 +292,11 @@ export default function Page() {
             return;
           }
         }
-      } else if (inputData.inputType === "spreadsheet" && inputData.spreadsheetImported) {
-        comments = inputData.spreadsheetData.map((row, index) => ({
+      } else if (inputType === "spreadsheet" && spreadsheetImported) {
+        comments = spreadsheetData.map((row, index) => ({
           id: row.id || `spreadsheet-${index + 1}`,
           comment: (row as unknown as Record<string, unknown>)[
-            inputData.selectedCommentColumn
+            selectedCommentColumn
           ] as string,
           source: row.source || null,
           url: row.url || null,
@@ -129,36 +313,109 @@ export default function Page() {
     }
 
     try {
-      const promptData = promptSettings.getPromptSettings();
-      
-      await createReport({
-        input: basicInfo.input,
-        question: basicInfo.question,
-        intro: basicInfo.intro,
+      const payload = {
+        input,
+        question,
+        intro,
         comments,
-        cluster: [clusterSettings.clusterLv1, clusterSettings.clusterLv2],
-        model: aiSettings.model,
-        workers: aiSettings.workers,
-        prompt: promptData,
-        is_pubcom: aiSettings.isPubcomMode,
-        inputType: inputData.inputType,
-      });
-      
+        cluster: [clusterLv1, clusterLv2],
+        model,
+        workers,
+        prompt: {
+          extraction,
+          initialLabelling,
+          mergeLabelling,
+          overview,
+        },
+        is_pubcom: isPubcomMode,
+      };
+
+      console.log("送信されるJSON:", payload);
+
+      const response = await fetch(
+        `${process.env.NEXT_PUBLIC_API_BASEPATH}/admin/reports`,
+        {
+          method: "POST",
+          headers: {
+            "x-api-key": process.env.NEXT_PUBLIC_ADMIN_API_KEY || "",
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify({
+            input,
+            question,
+            intro,
+            comments,
+            cluster: [clusterLv1, clusterLv2],
+            model,
+            workers,
+            prompt: {
+              extraction,
+              initialLabelling,
+              mergeLabelling,
+              overview,
+            },
+            is_pubcom: isPubcomMode,
+            inputType,
+          }),
+        },
+      );
+      if (!response.ok) {
+        throw new Error(response.statusText);
+      }
       toaster.create({
         duration: 5000,
         type: "success",
         title: "レポート作成を開始しました",
       });
-      
       router.replace("/");
     } catch (e) {
-      showErrorToast(toaster, e, "レポート作成に失敗しました");
+      console.error(e);
+      toaster.create({
+        duration: 5000,
+        type: "error",
+        title: "レポート作成に失敗しました",
+        description: "問題が解決しない場合は開発者に問い合わせください",
+      });
     } finally {
       setLoading(false);
     }
+  }
+  const handleTabValueChange = (details: { value: string }) => {
+    const newType = details.value as "file" | "spreadsheet";
+    setInputType(newType);
+
+    if (newType === "file" && csv) {
+      // CSVのカラムを再構築
+      parseCsv(csv).then((parsed) => {
+        if (parsed.length > 0) {
+          const columns = Object.keys(parsed[0]);
+          setCsvColumns(columns);
+          if (columns.includes("comment")) {
+            setSelectedCommentColumn("comment");
+          }
+          setRecommendedClusters(calculateRecommendedClusters(parsed.length));
+        }
+      });
+    } else if (newType === "spreadsheet" && spreadsheetData.length > 0) {
+      // スプレッドシートのカラムを再構築
+      const columns = Object.keys(spreadsheetData[0]);
+      setCsvColumns(columns);
+      if (columns.includes("comment")) {
+        setSelectedCommentColumn("comment");
+      }
+      const recommended = calculateRecommendedClusters(spreadsheetData.length);
+      setRecommendedClusters(recommended);
+      // 推奨クラスタ数を自動的に適用
+      setClusterLv1(recommended.lv1);
+      setClusterLv2(recommended.lv2);
+    } else {
+      // データがない場合はカラムリセット
+      setCsvColumns([]);
+      setSelectedCommentColumn("");
+      setRecommendedClusters(null);
+    }
   };
 
-  // メインコンポーネントのレンダリング
   return (
     <div className={"container"}>
       <Header />
@@ -167,23 +424,31 @@ export default function Page() {
           新しいレポートを作成する
         </Heading>
         <VStack gap={5}>
-          {/* 基本情報セクション */}
-          <BasicInfoSection
-            input={basicInfo.input}
-            question={basicInfo.question}
-            intro={basicInfo.intro}
-            isIdValid={basicInfo.isIdValid}
-            onIdChange={basicInfo.handleIdChange}
-            onQuestionChange={basicInfo.handleQuestionChange}
-            onIntroChange={basicInfo.handleIntroChange}
-          />
-
-          {/* 入力データセクション */}
+          <Field.Root>
+            <Field.Label>タイトル</Field.Label>
+            <Input
+              value={question}
+              onChange={(e) => setQuestion(e.target.value)}
+              placeholder="例：人類が人工知能を開発・展開する上で、最優先すべき課題は何でしょうか？"
+            />
+            <Field.HelperText>レポートのタイトルを記載します</Field.HelperText>
+          </Field.Root>
+          <Field.Root>
+            <Field.Label>調査概要</Field.Label>
+            <Input
+              value={intro}
+              onChange={(e) => setIntro(e.target.value)}
+              placeholder="例：このAI生成レポートは、パブリックコメントにおいて寄せられた意見に基づいています。"
+            />
+            <Field.HelperText>
+              コメントの集計期間や、コメントの収集元など、調査の概要を記載します
+            </Field.HelperText>
+          </Field.Root>
           <Field.Root>
             <Field.Label>入力データ</Field.Label>
             <Tabs.Root
               defaultValue="file"
-              value={inputData.inputType}
+              value={inputType}
               onValueChange={handleTabValueChange}
               variant="enclosed"
               width="100%"
@@ -197,68 +462,772 @@ export default function Page() {
               </Tabs.List>
 
               <Box p={4}>
-                {/* CSVファイルタブ */}
-                <CsvFileTab
-                  csv={inputData.csv}
-                  setCsv={inputData.setCsv}
-                  csvColumns={inputData.csvColumns}
-                  setCsvColumns={inputData.setCsvColumns}
-                  selectedCommentColumn={inputData.selectedCommentColumn}
-                  setSelectedCommentColumn={inputData.setSelectedCommentColumn}
-                  clusterSettings={clusterSettings}
-                />
+                <Tabs.Content value="file">
+                  <VStack alignItems="stretch" w="full">
+                    <Link
+                      href="/sample_comments.csv"
+                      download
+                      style={{
+                        display: "inline-flex",
+                        alignItems: "center",
+                        gap: "4px",
+                        marginLeft: "8px",
+                        textDecoration: "underline",
+                        fontSize: "0.8rem",
+                      }}
+                    >
+                      <DownloadIcon size={14} />
+                      サンプルCSVをダウンロード
+                    </Link>
+                    <FileUploadRoot
+                      w={"full"}
+                      alignItems="stretch"
+                      accept={["text/csv"]}
+                      inputProps={{ multiple: false }}
+                      onFileChange={async (e) => {
+                        const file = e.acceptedFiles[0];
+                        setCsv(file);
+                        if (file) {
+                          const parsed = await parseCsv(file);
+                          if (parsed.length > 0) {
+                            const columns = Object.keys(parsed[0]);
+                            setCsvColumns(columns);
+                            if (columns.includes("comment")) {
+                              setSelectedCommentColumn("comment");
+                            }
+                            const recommended = calculateRecommendedClusters(parsed.length);
+                            setRecommendedClusters(recommended);
+                            // 推奨クラスタ数を自動的に適用
+                            setClusterLv1(recommended.lv1);
+                            setClusterLv2(recommended.lv2);
+                          }
+                        }
+                      }}
+                    >
+                      <Box
+                        opacity={csv ? 0.5 : 1}
+                        pointerEvents={csv ? "none" : "auto"}
+                      >
+                        <FileUploadDropzone
+                          label="分析するコメントファイルを選択してください"
+                          description=".csv"
+                        />
+                      </Box>
+                      <FileUploadList
+                        clearable={true}
+                        onRemove={() => {
+                          setCsv(null);
+                          setCsvColumns([]);
+                          setSelectedCommentColumn("");
+                          setRecommendedClusters(null);
+                        }}
+                      />
+                    </FileUploadRoot>
+                    {csvColumns.length > 0 && (
+                      <Field.Root mt={4}>
+                        <Field.Label>コメントカラム選択</Field.Label>
+                        <NativeSelect.Root w={"60%"}>
+                          <NativeSelect.Field
+                            value={selectedCommentColumn}
+                            onChange={(e) =>
+                              setSelectedCommentColumn(e.target.value)
+                            }
+                          >
+                            <option value="">選択してください</option>
+                            {csvColumns.map((col) => (
+                              <option key={col} value={col}>
+                                {col}
+                              </option>
+                            ))}
+                          </NativeSelect.Field>
+                          <NativeSelect.Indicator />
+                        </NativeSelect.Root>
+                        <Field.HelperText>
+                          分析対象のコメントが含まれるカラムを選んでください（それ以外のカラムは無視されます）。
+                        </Field.HelperText>
+                      </Field.Root>
+                    )}
 
-                {/* スプレッドシートタブ */}
-                <SpreadsheetTab
-                  spreadsheetUrl={inputData.spreadsheetUrl}
-                  setSpreadsheetUrl={inputData.setSpreadsheetUrl}
-                  spreadsheetImported={inputData.spreadsheetImported}
-                  spreadsheetLoading={inputData.spreadsheetLoading}
-                  spreadsheetData={inputData.spreadsheetData}
-                  importedId={inputData.importedId}
-                  canImport={inputData.canImport}
-                  csvColumns={inputData.csvColumns}
-                  selectedCommentColumn={inputData.selectedCommentColumn}
-                  setSelectedCommentColumn={inputData.setSelectedCommentColumn}
-                  clusterSettings={clusterSettings}
-                  onImport={() => inputData.importSpreadsheet(basicInfo.input)}
-                  onClearData={inputData.clearSpreadsheetData}
-                />
+                    {recommendedClusters && (
+                      <Field.Root mt={4}>
+                        <Field.Label>クラスタ数設定</Field.Label>
+                        <HStack w={"100%"}>
+                          <Button
+                            onClick={() => {
+                              setClusterLv1(Math.max(2, clusterLv1 - 1));
+                            }}
+                            variant="outline"
+                          >
+                            -
+                          </Button>
+                          <Input
+                            type="number"
+                            value={clusterLv1.toString()}
+                            min={2}
+                            max={20}
+                            onChange={(e) => {
+                              const v = Number(e.target.value);
+                              if (!Number.isNaN(v)) {
+                                const newClusterLv1 = Math.max(2, Math.min(20, v));
+                                setClusterLv1(newClusterLv1);
+
+                                // 第一階層のクラスタ数 * 2 > 第二階層のクラスタ数の場合のみ、第二階層の値を更新
+                                const newClusterLv2 = newClusterLv1 * 2;
+                                if (newClusterLv2 > clusterLv2) {
+                                  setClusterLv2(newClusterLv2);
+                                  setAutoAdjusted(true);
+                                }
+                                
+                                // 推奨クラスタ数表示を更新（nullでない場合のみ）
+                                if (recommendedClusters) {
+                                  setRecommendedClusters({
+                                    lv1: newClusterLv1,
+                                    lv2: newClusterLv2 > clusterLv2 ? newClusterLv2 : clusterLv2
+                                  });
+                                }
+                              }
+                            }}
+                          />
+                          <Button
+                            onClick={() => {
+                              const newClusterLv1 = Math.min(20, clusterLv1 + 1);
+                              setClusterLv1(newClusterLv1);
+
+                              // 第一階層のクラスタ数 * 2 > 第二階層のクラスタ数の場合のみ、第二階層の値を更新
+                              const newClusterLv2 = newClusterLv1 * 2;
+                              if (newClusterLv2 > clusterLv2) {
+                                setClusterLv2(newClusterLv2);
+                                setAutoAdjusted(true);
+                              }
+                              
+                              // 推奨クラスタ数表示を更新（nullでない場合のみ）
+                              if (recommendedClusters) {
+                                setRecommendedClusters({
+                                  lv1: newClusterLv1,
+                                  lv2: newClusterLv2 > clusterLv2 ? newClusterLv2 : clusterLv2
+                                });
+                              }
+                            }}
+                            variant="outline"
+                          >
+                            +
+                          </Button>
+                          <ChevronRightIcon width="100px" />
+                          <Button
+                            onClick={() => {
+                              const newClusterLv2 = Math.max(2, clusterLv2 - 1);
+                              setClusterLv2(newClusterLv2);
+
+                              // 第二階層の値が第一階層の値の2倍未満の場合は自動調整
+                              if (newClusterLv2 < clusterLv1 * 2) {
+                                const adjustedValue = clusterLv1 * 2;
+                                setClusterLv2(adjustedValue);
+                                setAutoAdjusted(true);
+                                
+                                // 推奨クラスタ数表示を更新（nullでない場合のみ）
+                                if (recommendedClusters) {
+                                  setRecommendedClusters({
+                                    lv1: recommendedClusters.lv1,
+                                    lv2: adjustedValue
+                                  });
+                                }
+                              } else {
+                                setAutoAdjusted(false);
+                                
+                                // 推奨クラスタ数表示を更新（nullでない場合のみ）
+                                if (recommendedClusters) {
+                                  setRecommendedClusters({
+                                    lv1: recommendedClusters.lv1,
+                                    lv2: newClusterLv2
+                                  });
+                                }
+                              }
+                            }}
+                            variant="outline"
+                          >
+                            -
+                          </Button>
+                          <Input
+                            type="number"
+                            value={clusterLv2.toString()}
+                            min={2}
+                            max={1000}
+                            onChange={(e) => {
+                              // 入力中も最大値を制限
+                              const inputValue = e.target.value;
+                              if (inputValue === "") {
+                                return; // 空の入力は無視
+                              }
+
+                              const v = Number(inputValue);
+                              if (!Number.isNaN(v)) {
+                                // 最大値を1000に制限
+                                const limitedValue = Math.min(1000, v);
+                                setClusterLv2(limitedValue);
+                                
+                                // 推奨クラスタ数表示を更新（nullでない場合のみ）
+                                if (recommendedClusters) {
+                                  setRecommendedClusters({
+                                    lv1: recommendedClusters.lv1,
+                                    lv2: limitedValue
+                                  });
+                                }
+                              }
+                            }}
+                            onBlur={(e) => {
+                              // フォーカスが外れたときに値の検証と自動調整を行う
+                              const v = Number(e.target.value);
+                              if (!Number.isNaN(v)) {
+                                let newValue = Math.max(2, Math.min(1000, v));
+
+                                // 第二階層の値が第一階層の値の2倍未満の場合は自動調整
+                                if (newValue < clusterLv1 * 2) {
+                                  newValue = clusterLv1 * 2;
+                                  setClusterLv2(newValue);
+                                  setAutoAdjusted(true);
+                                } else {
+                                  setAutoAdjusted(false);
+                                }
+                                
+                                // 推奨クラスタ数表示を更新（nullでない場合のみ）
+                                if (recommendedClusters) {
+                                  setRecommendedClusters({
+                                    lv1: recommendedClusters.lv1,
+                                    lv2: newValue
+                                  });
+                                }
+                              }
+                            }}
+                          />
+                          <Button
+                            onClick={() => {
+                              const newClusterLv2 = Math.min(1000, clusterLv2 + 1);
+                              setClusterLv2(newClusterLv2);
+
+                              // 第二階層の値が第一階層の値の2倍未満の場合は自動調整
+                              if (newClusterLv2 < clusterLv1 * 2) {
+                                const adjustedValue = clusterLv1 * 2;
+                                setClusterLv2(adjustedValue);
+                                setAutoAdjusted(true);
+                                
+                                // 推奨クラスタ数表示を更新（nullでない場合のみ）
+                                if (recommendedClusters) {
+                                  setRecommendedClusters({
+                                    lv1: recommendedClusters.lv1,
+                                    lv2: adjustedValue
+                                  });
+                                }
+                              } else {
+                                setAutoAdjusted(false);
+                                
+                                // 推奨クラスタ数表示を更新（nullでない場合のみ）
+                                if (recommendedClusters) {
+                                  setRecommendedClusters({
+                                    lv1: recommendedClusters.lv1,
+                                    lv2: newClusterLv2
+                                  });
+                                }
+                              }
+                            }}
+                            variant="outline"
+                          >
+                            +
+                          </Button>
+                        </HStack>
+                        <Field.HelperText>
+                          階層ごとの意見グループ生成数です。初期値はコメント数に基づいた推奨クラスタ数です。
+                        </Field.HelperText>
+                        {autoAdjusted && (
+                          <Text color="orange.500" fontSize="sm" mt={2}>
+                            第2階層の意見グループ数が自動調整されました。第2階層の意見グループ数は第1階層の意見グループ数の2倍以上に設定してください。
+                          </Text>
+                        )}
+                      </Field.Root>
+                    )}
+                  </VStack>
+                </Tabs.Content>
+
+                <Tabs.Content value="spreadsheet">
+                  <VStack alignItems="stretch" w="full" gap={4}>
+                    <Field.Root>
+                      <Field.Label>スプレッドシートURL</Field.Label>
+                      <HStack
+                        w="full"
+                        flexDir={["column", "column", "row"]}
+                        alignItems={["stretch", "stretch", "center"]}
+                        gap={[2, 2, 4]}
+                      >
+                        <Input
+                          flex="1"
+                          value={spreadsheetUrl}
+                          onChange={(e) => setSpreadsheetUrl(e.target.value)}
+                          placeholder="https://docs.google.com/spreadsheets/d/xxxxxxxxxxxx/edit"
+                          disabled={spreadsheetImported} // インポート済みの場合はURL入力も無効化
+                        />
+                        <Button
+                          onClick={importSpreadsheet}
+                          loading={spreadsheetLoading}
+                          disabled={!canImport}
+                          whiteSpace="nowrap"
+                          width={["full", "full", "auto"]}
+                        >
+                          {spreadsheetImported ? "取得済み" : "データを取得"}
+                        </Button>
+                      </HStack>
+                      <Field.HelperText>
+                        公開されているGoogleスプレッドシートのURLを入力してください
+                        <br />
+                      </Field.HelperText>
+                      {csvColumns.length > 0 && (
+                        <Field.Root mt={4}>
+                          <Field.Label>コメントカラム選択</Field.Label>
+                          <NativeSelect.Root w={"60%"}>
+                            <NativeSelect.Field
+                              value={selectedCommentColumn}
+                              onChange={(e) =>
+                                setSelectedCommentColumn(e.target.value)
+                              }
+                            >
+                              <option value="">選択してください</option>
+                              {csvColumns.map((col) => (
+                                <option key={col} value={col}>
+                                  {col}
+                                </option>
+                              ))}
+                            </NativeSelect.Field>
+                            <NativeSelect.Indicator />
+                          </NativeSelect.Root>
+                          <Field.HelperText>
+                            分析対象のコメントが含まれるカラムを選んでください（それ以外のカラムは無視されます）。
+                          </Field.HelperText>
+                        </Field.Root>
+                      )}
+                      
+                      {recommendedClusters && (
+                        <Field.Root mt={4}>
+                          <Field.Label>クラスタ数設定</Field.Label>
+                          <HStack w={"100%"}>
+                            <Button
+                              onClick={() => {
+                                setClusterLv1(Math.max(2, clusterLv1 - 1));
+                              }}
+                              variant="outline"
+                            >
+                              -
+                            </Button>
+                            <Input
+                              type="number"
+                              value={clusterLv1.toString()}
+                              min={2}
+                              max={20}
+                              onChange={(e) => {
+                                const v = Number(e.target.value);
+                                if (!Number.isNaN(v)) {
+                                  const newClusterLv1 = Math.max(2, Math.min(20, v));
+                                  setClusterLv1(newClusterLv1);
+
+                                  // 第一階層のクラスタ数 * 2 > 第二階層のクラスタ数の場合のみ、第二階層の値を更新
+                                  const newClusterLv2 = newClusterLv1 * 2;
+                                  if (newClusterLv2 > clusterLv2) {
+                                    setClusterLv2(newClusterLv2);
+                                    setAutoAdjusted(true);
+                                  }
+                                  
+                                  // 推奨クラスタ数表示を更新（nullでない場合のみ）
+                                  if (recommendedClusters) {
+                                    setRecommendedClusters({
+                                      lv1: newClusterLv1,
+                                      lv2: newClusterLv2 > clusterLv2 ? newClusterLv2 : clusterLv2
+                                    });
+                                  }
+                                }
+                              }}
+                            />
+                            <Button
+                              onClick={() => {
+                                const newClusterLv1 = Math.min(20, clusterLv1 + 1);
+                                setClusterLv1(newClusterLv1);
+
+                                // 第一階層のクラスタ数 * 2 > 第二階層のクラスタ数の場合のみ、第二階層の値を更新
+                                const newClusterLv2 = newClusterLv1 * 2;
+                                if (newClusterLv2 > clusterLv2) {
+                                  setClusterLv2(newClusterLv2);
+                                  setAutoAdjusted(true);
+                                }
+                                
+                                // 推奨クラスタ数表示を更新（nullでない場合のみ）
+                                if (recommendedClusters) {
+                                  setRecommendedClusters({
+                                    lv1: newClusterLv1,
+                                    lv2: newClusterLv2 > clusterLv2 ? newClusterLv2 : clusterLv2
+                                  });
+                                }
+                              }}
+                              variant="outline"
+                            >
+                              +
+                            </Button>
+                            <ChevronRightIcon width="100px" />
+                            <Button
+                              onClick={() => {
+                                const newClusterLv2 = Math.max(2, clusterLv2 - 1);
+                                setClusterLv2(newClusterLv2);
+
+                                // 第二階層の値が第一階層の値の2倍未満の場合は自動調整
+                                if (newClusterLv2 < clusterLv1 * 2) {
+                                  const adjustedValue = clusterLv1 * 2;
+                                  setClusterLv2(adjustedValue);
+                                  setAutoAdjusted(true);
+                                  
+                                  // 推奨クラスタ数表示を更新（nullでない場合のみ）
+                                  if (recommendedClusters) {
+                                    setRecommendedClusters({
+                                      lv1: recommendedClusters.lv1,
+                                      lv2: adjustedValue
+                                    });
+                                  }
+                                } else {
+                                  setAutoAdjusted(false);
+                                  
+                                  // 推奨クラスタ数表示を更新（nullでない場合のみ）
+                                  if (recommendedClusters) {
+                                    setRecommendedClusters({
+                                      lv1: recommendedClusters.lv1,
+                                      lv2: newClusterLv2
+                                    });
+                                  }
+                                }
+                              }}
+                              variant="outline"
+                            >
+                              -
+                            </Button>
+                            <Input
+                              type="number"
+                              value={clusterLv2.toString()}
+                              min={2}
+                              max={1000}
+                              onChange={(e) => {
+                                // 入力中も最大値を制限
+                                const inputValue = e.target.value;
+                                if (inputValue === "") {
+                                  return; // 空の入力は無視
+                                }
+
+                                const v = Number(inputValue);
+                                if (!Number.isNaN(v)) {
+                                  // 最大値を1000に制限
+                                  const limitedValue = Math.min(1000, v);
+                                  setClusterLv2(limitedValue);
+                                  
+                                  // 推奨クラスタ数表示を更新（nullでない場合のみ）
+                                  if (recommendedClusters) {
+                                    setRecommendedClusters({
+                                      lv1: recommendedClusters.lv1,
+                                      lv2: limitedValue
+                                    });
+                                  }
+                                }
+                              }}
+                              onBlur={(e) => {
+                                // フォーカスが外れたときに値の検証と自動調整を行う
+                                const v = Number(e.target.value);
+                                if (!Number.isNaN(v)) {
+                                  let newValue = Math.max(2, Math.min(1000, v));
+
+                                  // 第二階層の値が第一階層の値の2倍未満の場合は自動調整
+                                  if (newValue < clusterLv1 * 2) {
+                                    newValue = clusterLv1 * 2;
+                                    setClusterLv2(newValue);
+                                    setAutoAdjusted(true);
+                                  } else {
+                                    setAutoAdjusted(false);
+                                  }
+                                  
+                                  // 推奨クラスタ数表示を更新（nullでない場合のみ）
+                                  if (recommendedClusters) {
+                                    setRecommendedClusters({
+                                      lv1: recommendedClusters.lv1,
+                                      lv2: newValue
+                                    });
+                                  }
+                                }
+                              }}
+                            />
+                            <Button
+                              onClick={() => {
+                                const newClusterLv2 = Math.min(1000, clusterLv2 + 1);
+                                setClusterLv2(newClusterLv2);
+
+                                // 第二階層の値が第一階層の値の2倍未満の場合は自動調整
+                                if (newClusterLv2 < clusterLv1 * 2) {
+                                  const adjustedValue = clusterLv1 * 2;
+                                  setClusterLv2(adjustedValue);
+                                  setAutoAdjusted(true);
+                                  
+                                  // 推奨クラスタ数表示を更新（nullでない場合のみ）
+                                  if (recommendedClusters) {
+                                    setRecommendedClusters({
+                                      lv1: recommendedClusters.lv1,
+                                      lv2: adjustedValue
+                                    });
+                                  }
+                                } else {
+                                  setAutoAdjusted(false);
+                                  
+                                  // 推奨クラスタ数表示を更新（nullでない場合のみ）
+                                  if (recommendedClusters) {
+                                    setRecommendedClusters({
+                                      lv1: recommendedClusters.lv1,
+                                      lv2: newClusterLv2
+                                    });
+                                  }
+                                }
+                              }}
+                              variant="outline"
+                            >
+                              +
+                            </Button>
+                          </HStack>
+                          <Field.HelperText>
+                            階層ごとの意見グループ生成数です。初期値はコメント数に基づいた推奨クラスタ数です。
+                          </Field.HelperText>
+                          {autoAdjusted && (
+                            <Text color="orange.500" fontSize="sm" mt={2}>
+                              第2階層の意見グループ数が自動調整されました。第2階層の意見グループ数は第1階層の意見グループ数の2倍以上に設定してください。
+                            </Text>
+                          )}
+                        </Field.Root>
+                      )}
+                    </Field.Root>
+                    {spreadsheetImported && (
+                      <Text color="green.500" fontSize="sm">
+                        スプレッドシートのデータ {spreadsheetData.length}{" "}
+                        件を取得しました
+                      </Text>
+                    )}
+
+                    {/* スプレッドシートデータ再取得のためのボタンを追加 */}
+                    {spreadsheetImported && (
+                      <Button
+                        onClick={async () => {
+                          try {
+                            setSpreadsheetLoading(true);
+                            // サーバー側のデータを削除するAPI呼び出し - インポート時のIDを使用
+                            const response = await fetch(
+                              `${process.env.NEXT_PUBLIC_API_BASEPATH}/admin/inputs/${importedId}`,
+                              {
+                                method: "DELETE",
+                                headers: {
+                                  "x-api-key":
+                                    process.env.NEXT_PUBLIC_ADMIN_API_KEY || "",
+                                },
+                              },
+                            );
+
+                            if (!response.ok) {
+                              throw new Error("データのクリアに失敗しました");
+                            }
+
+                            toaster.create({
+                              type: "success",
+                              title: "成功",
+                              description: "データをクリアしました",
+                            });
+                          } catch (e) {
+                            console.error(e);
+                            toaster.create({
+                              type: "warning",
+                              title: "警告",
+                              description:
+                                "サーバー側のデータクリアに失敗しましたが、入力をリセットしました",
+                            });
+                          } finally {
+                            // UIの状態をリセット
+                            setSpreadsheetImported(false);
+                            setSpreadsheetData([]);
+                            setSpreadsheetLoading(false);
+                            setImportedId("");
+                            setSpreadsheetUrl("");
+                            setSelectedCommentColumn("");
+                            setCsvColumns([]);
+                          }
+                        }}
+                        colorScheme="red"
+                        variant="outline"
+                        size="sm"
+                        loading={spreadsheetLoading}
+                        loadingText="クリア中..."
+                      >
+                        データをクリアして再入力
+                      </Button>
+                    )}
+                  </VStack>
+                </Tabs.Content>
               </Box>
             </Tabs.Root>
           </Field.Root>
-
-          {/* AI詳細設定ボタン */}
           <HStack justify={"flex-end"} w={"full"}>
             <Button onClick={onToggle} variant={"outline"} w={"200px"}>
               AI詳細設定 (オプション)
             </Button>
           </HStack>
-
-          {/* AI詳細設定セクション */}
           <Presence present={open} w={"full"}>
-            <AISettingsSection
-              model={aiSettings.model}
-              workers={aiSettings.workers}
-              isPubcomMode={aiSettings.isPubcomMode}
-              onModelChange={aiSettings.handleModelChange}
-              onWorkersChange={(e) => {
-                const v = Number(e.target.value);
-                if (!Number.isNaN(v)) {
-                  aiSettings.handleWorkersChange(v);
-                }
-              }}
-              onIncreaseWorkers={aiSettings.increaseWorkers}
-              onDecreaseWorkers={aiSettings.decreaseWorkers}
-              onPubcomModeChange={aiSettings.handlePubcomModeChange}
-              getModelDescription={aiSettings.getModelDescription}
-            />
+            <VStack gap={10}>
+              <Field.Root>
+                <Field.Label>ID</Field.Label>
+                <Input
+                  w={"40%"}
+                  value={input}
+                  onChange={handleIdChange}
+                  placeholder="例：example"
+                  aria-invalid={!isIdValid}
+                  borderColor={!isIdValid ? "red.300" : undefined}
+                  _hover={!isIdValid ? { borderColor: "red.400" } : undefined}
+                />
+                {!isIdValid && (
+                  <Text color="red.500" fontSize="sm" mt={1}>
+                    IDは英小文字、数字、ハイフンのみ使用できます
+                  </Text>
+                )}
+                <Field.HelperText>
+                  英字小文字と数字とハイフンのみ(URLで利用されます)
+                </Field.HelperText>
+              </Field.Root>
+              <Field.Root>
+                <Checkbox
+                  checked={isPubcomMode}
+                  onCheckedChange={(details) => {
+                    const { checked } = details;
+                    if (checked === "indeterminate") return;
+                    setIsPubcomMode(checked);
+                  }}
+                >
+                  csv出力モード
+                </Checkbox>
+                <Field.HelperText>
+                  元のコメントと要約された意見をCSV形式で出力します。完成したCSVファイルはレポート一覧ページからダウンロードできます。
+                </Field.HelperText>
+              </Field.Root>
+              <Field.Root>
+                <Field.Label>並列実行数</Field.Label>
+                <HStack>
+                  <Button
+                    onClick={() => {
+                      setWorkers(Math.max(1, workers - 1));
+                    }}
+                    variant="outline"
+                  >
+                    -
+                  </Button>
+                  <Input
+                    type="number"
+                    value={workers.toString()}
+                    min={1}
+                    max={100}
+                    onChange={(e) => {
+                      const v = Number(e.target.value);
+                      if (!Number.isNaN(v)) {
+                        setWorkers(Math.max(1, Math.min(100, v)));
+                      }
+                    }}
+                  />
+                  <Button
+                    onClick={() => {
+                      setWorkers(Math.min(100, workers + 1));
+                    }}
+                    variant="outline"
+                  >
+                    +
+                  </Button>
+                </HStack>
+                <Field.HelperText>
+                  OpenAI
+                  APIの並列実行数です。値を大きくすることでレポート出力が速くなりますが、OpenAIアカウントのTierによってはレートリミットの上限に到達し、レポート出力が失敗する可能性があります。
+                </Field.HelperText>
+              </Field.Root>
+              <Field.Root>
+                <Field.Label>AIモデル</Field.Label>
+                <NativeSelect.Root w={"40%"}>
+                  <NativeSelect.Field
+                    value={model}
+                    onChange={(e) => setModel(e.target.value)}
+                  >
+                    <option value={"gpt-4o-mini"}>OpenAI GPT-4o mini</option>
+                    <option value={"gpt-4o"}>OpenAI GPT-4o</option>
+                    <option value={"o3-mini"}>OpenAI o3-mini</option>
+                  </NativeSelect.Field>
+                  <NativeSelect.Indicator />
+                </NativeSelect.Root>
+                <Field.HelperText>
+                  {model === "gpt-4o-mini" &&
+                    "GPT-4o mini：最も安価に利用できるモデルです。価格の詳細はOpenAIが公開しているAPI料金のページをご参照ください。"}
+                  {model === "gpt-4o" &&
+                    "GPT-4o：gpt-4o-miniと比較して高性能なモデルです。性能は高くなりますが、gpt-4o-miniと比較してOpenAI APIの料金は高くなります。"}
+                  {model === "o3-mini" &&
+                    "o3-mini：gpt-4oよりも高度な推論能力を備えたモデルです。性能はより高くなりますが、gpt-4oと比較してOpenAI APIの料金は高くなります。"}
+                </Field.HelperText>
+              </Field.Root>
+              <Field.Root>
+                <Field.Label>抽出プロンプト</Field.Label>
+                <Textarea
+                  h={"150px"}
+                  value={extraction}
+                  onChange={(e) => setExtraction(e.target.value)}
+                />
+                <Field.HelperText>
+                  AIに提示する抽出プロンプトです(通常は変更不要です)
+                </Field.HelperText>
+              </Field.Root>
+              <Field.Root>
+                <Field.Label>初期ラベリングプロンプト</Field.Label>
+                <Textarea
+                  h={"150px"}
+                  value={initialLabelling}
+                  onChange={(e) => setInitialLabelling(e.target.value)}
+                />
+                <Field.HelperText>
+                  AIに提示する初期ラベリングプロンプトです(通常は変更不要です)
+                </Field.HelperText>
+              </Field.Root>
+              <Field.Root>
+                <Field.Label>統合ラベリングプロンプト</Field.Label>
+                <Textarea
+                  h={"150px"}
+                  value={mergeLabelling}
+                  onChange={(e) => setMergeLabelling(e.target.value)}
+                />
+                <Field.HelperText>
+                  AIに提示する統合ラベリングプロンプトです(通常は変更不要です)
+                </Field.HelperText>
+              </Field.Root>
+              <Field.Root>
+                <Field.Label>要約プロンプト</Field.Label>
+                <Textarea
+                  h={"150px"}
+                  value={overview}
+                  onChange={(e) => setOverview(e.target.value)}
+                />
+                <Field.HelperText>
+                  AIに提示する要約プロンプトです(通常は変更不要です)
+                </Field.HelperText>
+              </Field.Root>
+            </VStack>
           </Presence>
-
-          {/* 警告メッセージ */}
-          <WarningSection />
-
-          {/* 送信ボタン */}
+          <Stack gap="4" width="full">
+            <Alert.Root status="warning">
+              <Alert.Indicator />
+                <Alert.Title>
+                  レポートの作成には数分から数十分程度の時間がかかります
+                </Alert.Title>
+            </Alert.Root>
+            <Alert.Root status="warning">
+              <Alert.Indicator />
+                <Alert.Title>
+                  作成が完了したレポートが一覧画面に反映されるまで5分程度かかります
+                </Alert.Title>
+            </Alert.Root>
+          </Stack>
           <Button
             mt={10}
             className={"gradientBg shadow"}
